+++ BoxOffice/Assets.xcassets/box_office_sample.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "data" : [
+    {
+      "filename" : "box_office_sample.json",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
