+++ BoxOffice/Model/BoxOffice.swift
@@ -1,0 +1,12 @@
+//
+//  BoxOffice.swift
+//  BoxOffice
+//
+//  Created by Jinah Park on 2023/03/20.
+//
+
+import Foundation
+
+struct BoxOffice: Codable {
+    let boxOfficeResult: BoxOfficeResult
+}
